+++ others/convertES/basic.js
@@ -6,7 +6,7 @@
 module.exports = (str, goal = "ES6") => {
     if (typeof str != "string" || !str) return;
     return goal == "ES6" ? str.replace(
-        /import\s+(?:(?:([A-z_$][\w$]*)|\*\s+as\s+([A-z_$][\w$]*)|({\s*[A-z_$][\w$]*(?:\s+as\s+[A-z_$][\w$]*)?\s*(?:,\s*[A-z_$][\w$]*(?:\s+as\s+[A-z_$][\w$]*)?\s*)?}))\s+from\s+(("|'|`).*\5)|(("|'|`).*\7))/g,
+        /import\s+(?:(?:([A-Za-z_$][\w$]*)|\*\s+as\s+([A-Za-z_$][\w$]*)|({\s*[A-Za-z_$][\w$]*(?:\s+as\s+[A-Za-z_$][\w$]*)?\s*(?:,\s*[A-Za-z_$][\w$]*(?:\s+as\s+[A-Za-z_$][\w$]*)?\s*)?}))\s+from\s+(("|'|`).*\5)|(("|'|`).*\7))/g,
         (_, ...a) => a[0] || a[1] || a[2] ? `const ${a[0] || a[1] || a[2]} = require(${a[3]})` : a[5] && `require(${a[5]})`
-    ).replace(/([A-z_$][\w$]*)\s+as\s+([A-z_$][\w$]*)/g, "$1: $2") : null;
+    ).replace(/([A-Za-z_$][\w$]*)\s+as\s+([A-Za-z_$][\w$]*)/g, "$1: $2") : null;
 };